CRUDAPP/CRUDAPP/requirements.txt
apiview==1.3.25
certifi==2020.4.5.2
cycler==0.10.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1.5
django-import-export==2.2.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
jdcal==1.4.1
kiwisolver==1.2.0
MarkupPy==1.14
matplotlib==3.1.3
mysqlclient==1.4.6
numpy==1.18.1
odfpy==1.4.1
openpyxl==3.0.4
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
tablib==2.0.0
xlrd==1.2.0
xlwt==1.3.0
